[fix]: pandas.Index.to_numpy() method's arguments not properly typed
After cloning a fresh copy of Pandas.
After generating new stubs using stubtest.
After looking for all the type signature for to_numpy method.

There was nothing to update in the following files:
- ./_libs/tslibs/nattype.pyi
- ./_libs/tslibs/timedeltas.pyi

There was modification to do in the following files:
- ./core/arrays/masked.pyi
- ./core/arrays/numpy_.pyi
- ./core/arrays/base.pyi
- /core/frame.pyi
- ./core/base.pyi

The type signature was removed from the following file:
- ./core/indexes/base.pyi

Refs #237